Refactor dialog example
