+++ examples/dialog_async/screenshot.png
